make permission checker extendable
